+++ src/main/webapp/view_user.jsp
@@ -8,6 +8,13 @@
 
 <%@ include file="includes/header.jsp"%>
 <%@ include file="includes/navbar_users.jsp"%>
+
+<script>
+    function confirmDelete() {
+        return confirm("¿Estás seguro de que quieres eliminar este usuario?");
+    }
+</script>
+
 <div class="album py-5 bg-body-tertiary">
     <div class="container">
         <div class="row row-cols-1 row-cols-sm-2 row-cols-md-3 g-3">
@@ -47,7 +54,7 @@
             } else if (role.equals("admin")) {
             %>
             <a href="edit_user.jsp?user_id=<%=user.getId()%>" class="btn btn-sm btn-warning">Edit</a>
-            <a href="delete_user?user_id=<%=user.getId()%>" class="btn btn-sm btn-danger">Eliminar</a>
+            <a onclick="return confirmDelete()" href="delete_user?user_id=<%=user.getId()%>" class="btn btn-sm btn-danger">Eliminar</a>
             <%
             } else {
             %>
